Merge commit 'f904b95ef075e652023a3832799c0f96fa3d3200' into upgrade01

+++ src/gallium/drivers/zink/zink_fence.h
@@ -51,7 +51,7 @@ struct zink_tc_fence {
 
 struct zink_fence {
    VkFence fence;
-   uint32_t batch_id;
+   uint64_t batch_id;
    bool submitted;
    bool completed;
 };
